// src/app/BusinessModel/Basketball/Basketball-score-info-type-enum.ts
export enum BasketballScoreInfoTypeEnum {
    firstQuarter = 29,
    secondQuarter = 30,
    thirdQuarter = 31,
    forthQuarter = 32,
    Set3 = 13,
    Set4 = 14,
    Set5 = 15,
    CP = 16,
    CS = 9,
    CFS = 10,
    Finished = 5,
    FinishedOverTime = 23,
    FinishedAfterPenalty = 24,
    OverTime = 28
}